requirements-dev.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-playwright>=0.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0